Loop association between Loop Constructs and do-loops
```
!$omp parallel do
do i=1, N
  ...
enddo
[!$omp end parallel do]
```
Here LoopDirective and EndLoopDirective are not wrapped in the same class.
Ideally it would be {LoopDirective, ClauseList, DoConstruct, EndLoopDirective}.
However, to allow different forms of do-loops, OpenMPLoopConstruct is actually
{LoopDirective, ClauseList} and the loop association checks are deferred to
Semantics.

In the example, do-loop needs to immediately follow the LoopDirective and
EndLoopDirective needs to immediately follow the ENDDO (DoConstruct). Here
the EndLoopdirective is optional because the scope of PARALLEL region is
determined by the scope of the associated do-loop. Unlike BlockDirective,
we pop the context upon the LEAVE of DoConstruct or on the point of failing
to find the associated do-loop.

Msg examples:
```
omp-clause-validity01.f90:220:15: error: do-loop is expected after the PARALLEL DO directive
          !$omp parallel do shared(b)
                ^^^^^^^^^^^
omp-clause-validity01.f90:224:12: error: The END PARALLEL DO must follow the do-loop associated with the loop construct
       !$omp end parallel do
             ^^^^^^^^^^^^^^^
```